Assign from buyer groups --> buyer prompt
When removing all of a buyers user groups from a price, admins will be prompted
if they want to assign the price at the buyer level instead. If yes, the buyer assignment
will be added and all usergroup assignments will be removed. If no, the removeBuyerAssignment
function is called which will warn them if this is the last buyer that the price will
be removed from the product.