Cleaning up app.controller.js
New Stuff:
- `ocStateLoading` service for managing state loading promises
- `ocTouch.config.js` for providing a boolean constant
`ocIsTouchDevice`
- `application.stateLoading` is now a function where you pass in
the name of the state that has the ui-view you want display a
loading indicator on when transitioning to sibling states.
  example: on `base.tpl.html`
  cg-busy="application.stateLoading('base')"

Changes
- Removed the $rootScope broadcasts in `oc-interceptors.config.js`
and used $injector to perform what used to be done when those
events were caught in app.controller.js
- Moved the $state event handlers to the app.run.js file

Temporary
- Added a message to the homepage that will tell you the value of
ocIsTouchDevice (helpful when debugging touch issues) from
an actual device.