Enhancement: ocCatalogs.Delete
The catalog delete service method now checks to see if buyers are
currently using the catalog to be deleted as a default catalog. If so,
I present the user with a modal for reassigning those buyer companies
to a new default catalog. The catalog is not deleted until all of the
buyers have been reassigned to a new catalog.

The flow is:
1. Confirm Delete Catalog
2. Check Buyers
   - if exist, Continue
   - if none, Delete catalog
3. Select new catalog
4. Assign selected catalog (vap, vac = true)
5. Patch all buyer's default catalog ID
6. Delete old catalog