F51-546 - Refactor oc-file-upload
The oc-file-upload directive is now responsible for opening the modal,
displaying the current image, and maintaining the xp properties tied
to the model.

When multiple is true, we will display the files in a gallery-like UI
where they can be dragged and dropped to change the order of the xp
array (draggable must be true).

For single file uploads, the file object will be stored directly on XP
For multiple file upload, the file object will be stored in an array
on XP

The $ocFiles provider now has the default values that are later merged
with the options passed into the directive.

ocFileReader was deemed unnecessary when uploading files to AWS, but
we will keep it around just in case OrderCloud.io supports files in
the future.

The file upload modal now contains all of the functionality for
uploading/validating files. This may later be updated to include a way
for users to browse existing files in their S3 bucket for reuse.